+++ src/main/java/com/example/communityProject/service/PostService.java
@@ -1,16 +1,19 @@
 package com.example.communityProject.service;
 
 import com.example.communityProject.dto.PostDto;
-import com.example.communityProject.dto.UserDto;
 import com.example.communityProject.entity.Image;
 import com.example.communityProject.entity.Post;
 import com.example.communityProject.entity.User;
 import com.example.communityProject.repository.*;
+import com.example.communityProject.security.JwtUtil;
+import jakarta.persistence.EntityNotFoundException;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.io.IOException;
 import java.time.LocalDateTime;
@@ -32,6 +35,8 @@ public class PostService {
     private ImageRepository imageRepository;
     @Autowired
     private ImageService imageService;
+    @Autowired
+    private JwtUtil jwtUtil;
 
     public List<PostDto> getPostList() {
         return postRepository.findAll()
@@ -56,10 +61,17 @@ public class PostService {
     }
 
     @Transactional
-    public PostDto updatePost(Long id, PostDto dto) {
+    public PostDto updatePost(Long id, PostDto dto, String token) {
+        Long userId = jwtUtil.getUserIdFromToken(token);
         // 타깃 조회하기
         Post target = postRepository.findById(id)
-                .orElseThrow(() -> new IllegalArgumentException("게시글 수정 실패, 대상 게시글이 없습니다."));
+                .orElseThrow(() -> new EntityNotFoundException("게시글 수정 실패, 대상 게시글이 없습니다."));
+        // 현재 로그인한 사용자가 작성자인지 확인
+        log.info("게시글 작성자: "+target.getUser().getId()+"로그인한 사용자: "+userId);
+        if (!target.getUser().getId().equals(userId)) {
+            throw new IllegalArgumentException("수정 권한이 없습니다.");
+        }
+
         // 업데이트하기
         target.patch(dto); // 기존 데이터에 새 데이터 붙이기
         Post updated = postRepository.save(target);
@@ -67,10 +79,29 @@ public class PostService {
     }
 
     @Transactional
-    public PostDto deletePost(Long id) {
+    public PostDto updatePostView(Long id, PostDto dto) {
+        // 타깃 조회하기
+        Post target = postRepository.findById(id)
+                .orElseThrow(() -> new EntityNotFoundException("게시글 조회수 증가 실패, 대상 게시글이 없습니다."));
+
+        // 업데이트하기
+        target.patch(dto); // 기존 데이터에 새 데이터 붙이기
+        Post updated = postRepository.save(target);
+        return PostDto.createPostDto(updated);
+    }
+
+
+    @Transactional
+    public PostDto deletePost(Long id, String token) {
+        Long userId = jwtUtil.getUserIdFromToken(token);
         // 대상 찾기
         Post target = postRepository.findById(id)
-                .orElseThrow(() -> new IllegalArgumentException("게시글 삭제 실패, 대상 게시글이 없습니다."));
+                .orElseThrow(() -> new EntityNotFoundException("게시글 삭제 실패, 대상 게시글이 없습니다."));
+
+        // 작성자와 현재 로그인한 사용자 비교
+        if (!target.getUser().getId().equals(userId)) {
+            throw new IllegalArgumentException("수정 권한이 없습니다.");
+        }
         // 관련 댓글 삭제
         commentRepository.deleteByPost_Id(id);
         // 관련 좋아요 삭제
